Add tracking and verification for each task
Added tracking for each task to verify it has been complete. So far, it checks if you're walking, how much you've walked, etc, same for biking, and for cleaning, instead of an AI model, which was the original plan, we check for where the picture was taken, as well as the time period it was taken in. Successfully doing these tasks gives climate coins, and failure to do them does not.